You can now submit a team.
It will check to see if the team already exists. If not, then it will add a new one. Then it will check to see if the URL was already added into the teams clues found. If not, then it adds it in. The leaderboard is also counting how many clues each team has found. Sadly, no good sort of greatest to lowest right now.